models & in memory db

+++ src/main/java/store/services/StoreService.java
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Service;
 public class StoreService {
 
 
-    public Double getBasketContent() {
+    public Integer getBasketContent() {
+
         return null;
     }
 }
